告警修改

+++ interfaces/inner_api/base/src/zchar_wrapper.cpp
@@ -128,7 +128,7 @@ zchar Char::GetChar(const std::string &str,
     int now = 0;
     while (*p && p < end) {
         zchar unicode = GetCharInternal((unsigned char *)p, bsize);
-        if (bsize == 0 || now + bsize > size) {
+        if (bsize == 0 || now + bsize > static_cast<int>(size)) {
             break;
         }
